1.1.4.8-Leisure Upgrade (Mandatory for TestNet)
- Added IsValidForPayment(Spork14) to masternodelist full (this shows if
a watchdog_expired sanc will be voted on)
- Added exec ipfsget hash filename - this allows a biblepay node to
download an ipfs file
- Added exec ipfslist - this is a report of hosted hashes from biblepay
transactions in last 7 days
- Added exec ipfsquality - This is a report of Sancs who are enabled
with their corresponding ipfs quality levels
- Added ability for a sanc to check another sancs IPFS health
- Added spork to support no team requirement in PODC and spork to
support team blacklisting
- Added icon for IPFS attachment (currently two sheets of paper, maybe
paperclip later)
- In transactiondescriptiondialog, make show attachment invisible when
no attachment, make links more intuitive and add backup link
- Fixed the filepicker three forward slashes issue
- Adjusted miner logic to prevent log-spam, and fixed the root cause of
the fork (it was creation of a block with the fMining flag on)
- Added PODS Sanc-side verification process; this penalizes sancs not
hosting files on IPFS; with a theoretical reward of :
hosted_file_percent * reward, achieved by voting for verified sancs

+++ src/wallet/crypter.cpp
@@ -173,13 +173,13 @@ void PrintStratisKeyDebugInfo()
 	std::string sKey1(reinterpret_cast<char*>(chKeyBiblePay));
 	std::string sIV1(reinterpret_cast<char*>(chIVBiblePay));
 	std::string sRow = "chKeyBiblePay: ";
-	for (int i = 0; i < sKey1.length(); i++)
+	for (int i = 0; i < (int)sKey1.length(); i++)
 	{
 		int ichar = (int)sKey1[i];
 		sRow += RoundToString(ichar,0) + ",";
 	}
 	sRow += "  chIVBiblePay: ";
-	for (int i = 0; i < sIV1.length(); i++)
+	for (int i = 0; i < (int)sIV1.length(); i++)
 	{
 		int ichar = (int)sIV1[i];
 		sRow += RoundToString(ichar,0) + ",";
